μEmacsv1.12.0:
C23 compile-time safety pass across entire codebase + code hygiene cleanup.

C23 COMPILE-TIME SHIFT (120+ files):
- [[nodiscard]] on all fallible functions (allocations, lookups, parsing)
- constexpr for compile-time constants replacing #define and const
- static_assert for struct layout and buffer size validation
- nullptr replacing NULL throughout (100% migration)
- restrict on non-aliasing pointer parameters
- Designated initializers for struct clarity
- Native bool (not typedef int)
- -Wall -Wextra -Wpedantic -Wconversion -Wsign-conversion zero warnings
- -fstack-protector-strong -fcf-protection FORTIFY_SOURCE=3 hardening
- ASan + UBSan in Debug builds

FIXES:
- Replace strncpy with snprintf in UEP extension system (C23 best practice)
  ext_host.c: resp.response copy used strncpy with manual null-termination
  extension_api.c: config entry key copy hit same pattern in set_int/set_bool/set_string
  snprintf always null-terminates, eliminates -Wstringop-truncation warnings

CLEANUP:
- Remove dead keyword_cmp() from lexer.c (strcmp comparator for bsearch never called)
- Fix misleading section header "binary search on sorted arrays" to "Keyword Lookup"
- Rename curses.c -> ansi.c (pure POSIX termios + ANSI escapes, zero curses dependency)
- Update all references: CMakeLists.txt, efunc.h, functions.h, display_ops.c, CLAUDE.md

Files Changed:
- All src/**/*.c, include/**/*.h, tests/**/*.c (C23 attribute + type safety pass)
- src/uep/ext_host.c (strncpy -> snprintf)
- src/uep/extension_api.c (strncpy -> snprintf)
- src/syntax/lexer.c (remove dead function, fix comment)
- src/terminal/drivers/curses.c -> src/terminal/drivers/ansi.c (rename)
- CMakeLists.txt (version bump 1.11.2 -> 1.12.0, source rename, compiler flags)
- include/internal/efunc.h (comment update)
- include/μemacs/functions.h (comment update)
- src/terminal/display_ops.c (comment update)
- .claude/CLAUDE.md (comment update)

+++ include/μemacs/utf8_optimized.h
@@ -108,8 +108,8 @@ static inline int utf8_display_width_cached(const unsigned char *utf8_str, int b
     
     /* Update cache atomically */
     entry->codepoint = cache_key;
-    entry->display_width = width;
-    entry->byte_length = byte_len;
+    entry->display_width = (uint8_t)width;
+    entry->byte_length = (uint8_t)byte_len;
     atomic_store_explicit(&entry->valid, true, memory_order_release);
     
     return width;
